fix: Infinite scroll
