+++ src/ducks/transactions/TransactionHeader.jsx
@@ -8,8 +8,18 @@ import { Breadcrumb } from 'components/Breadcrumb'
 import { ConnectedSelectDates } from 'components/SelectDates'
 import { AccountSwitch } from 'ducks/account'
 import TransactionSelectDates from 'ducks/transactions/TransactionSelectDates'
+import { TransactionTableHead } from './Transactions'
+import flag from 'cozy-flags'
+import HistoryChart from 'ducks/balance/HistoryChart'
 
 import styles from './TransactionsPage.styl'
+
+const historyChartMargin = {
+  top: 10,
+  bottom: 10,
+  left: 16,
+  right: 16
+}
 
 class TransactionHeader extends Component {
   isSubcategory = () => {
@@ -75,12 +85,40 @@ class TransactionHeader extends Component {
     )
   }
 
+  displayTableHead() {
+    const { t, breakpoints, router } = this.props
+
+    if (breakpoints.isMobile) {
+      return null
+    }
+
+    return (
+      <TransactionTableHead
+        mainColumnTitle={t(
+          router.params.subcategoryName
+            ? 'Categories.headers.movements'
+            : 'Transactions.header.description'
+        )}
+      />
+    )
+  }
+
+  displayBalanceHistory() {
+    if (!flag('balance-history')) {
+      return
+    }
+
+    return <HistoryChart margin={historyChartMargin} />
+  }
+
   render() {
     return (
       <div className={styles.TransactionPage__top}>
+        {this.displayBalanceHistory()}
         {this.displayAccountSwitch()}
         {this.displaySelectDates()}
         {this.displayBreadcrumb()}
+        {this.displayTableHead()}
       </div>
     )
   }
